requirements.txt
torch>=2.2.0
PyYAML>=6.0.1
pytest>=8.0.0
tokenizers>=0.15.2
datasets>=2.19.0
tqdm>=4.66.0
